modified:   package-lock.json 	modified:   package.json 	renamed:    src/routes/__error.svelte -> src/routes/+error.svelte 	renamed:    src/routes/__layout.svelte -> src/routes/+layout.svelte 	renamed:    src/routes/index.js -> src/routes/+page.server.js 	renamed:    src/routes/index.svelte -> src/routes/+page.svelte 	renamed:    src/routes/404.svelte -> src/routes/404/+page.svelte 	renamed:    src/routes/blog/index.js -> src/routes/blog/+page.server.js 	renamed:    src/routes/blog/index.svelte -> src/routes/blog/+page.svelte 	renamed:    src/routes/blog/welcome.md -> src/routes/blog/welcome/+page.md 	renamed:    src/routes/contact/index.md -> src/routes/contact/+page.md 	renamed:    src/routes/file/index.js -> src/routes/file/+page.server.js 	renamed:    src/routes/file/index.svelte -> src/routes/file/+page.svelte 	renamed:    src/routes/file/test.md -> src/routes/file/test/+page.md 	renamed:    src/routes/privacy-policy/index.md -> src/routes/privacy-policy/+page.md 	renamed:    src/routes/profiles/elclark.md -> src/routes/profiles/elclark/+page.md 	renamed:    src/routes/terms-and-conditions/index.md -> src/routes/terms-and-conditions/+page.md
