+++ src/routes/file/+page.svelte
@@ -1,4 +1,6 @@
 <script>
+	throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
 	import moment from 'moment'
 	import { page } from '$app/stores'
 	import bytesToSize from '$lib/bytesToSize'
